requirements.txt
accelerate==1.2.1
annotated-types==0.7.0
anyio==4.9.0
bitsandbytes==0.45.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
distro==1.9.0
filelock==3.18.0
fsspec==2025.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.70.0
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.32.0
idna==3.10
jinja2==3.1.6
jiter==0.10.0
markupsafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
openai==1.58.1
packaging==25.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.5
pydantic-core==2.33.2
pyparsing==3.2.3
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
safetensors==0.5.3
scipy==1.15.1
setuptools==80.8.0
sniffio==1.3.1
sympy==1.13.1
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.48.0
triton==3.1.0
typing-extensions==4.13.2
typing-inspection==0.4.1
uritemplate==4.1.1
urllib3==2.4.0